Added content to the home page (index.html)
I carried over the text content from the home page of
https://fdlhorseproject.wixsite.com/home to the home page of this
website.

I also added the image depicted on the original site of a drawing of a
horse. This image currently exists to the right of the text content of
the page.

Overall formatting could be improved upon.

The footer could be improved upon. Right now it simply shares the
copyright information listed on the original site.

 On branch main
 Changes to be committed:
	modified:   css/styles.css
	modified:   html/index.html
	new file:   images/horse_art.avif