Additional touch-ups to create consistency
Several touch-ups were made throughout the site to create more consistency
between each page.

Some links were added (namely in the footer of each
page).

The footer was corrected on each page such that it appears uniformly across
each page.

Some CSS was generalized to better fit the site as a whole. This was done
by changing "trophy-content" and "trophy-section" to "main-content" and
"main-section" respectively, as they were used on multiple pages unrelated
to the trophy page, where this styling originated from.

"formsHorseLogo.jpg" was already committed, but was potentially deleted or
lost. It has been readded to the project if needed.

 On branch main
 Changes to be committed:
	modified:   css/styles.css
	modified:   html/Trophies.html
	modified:   html/board.html
	modified:   html/calendar.html
	modified:   html/constitution.html
	modified:   html/contactUs.html
	modified:   html/forms.html
	modified:   html/index.html
	modified:   html/photo.html
	new file:   images/formsHorseLogo.jpg